Moved EmptyChunkGenerator into its own separate class file
