+++ src/main/java/club/bastonbolado/voidgenerator/EmptyChunkGenerator.java
@@ -1,0 +1,30 @@
+package club.bastonbolado.voidgenerator;
+
+import org.bukkit.Location;
+import org.bukkit.Material;
+import org.bukkit.World;
+import org.bukkit.generator.ChunkGenerator;
+
+import java.util.Random;
+
+public class EmptyChunkGenerator extends ChunkGenerator {
+
+    @Override
+    public ChunkData generateChunkData(World world, Random random, int x, int z, BiomeGrid biome) {
+        return createChunkData(world);
+    }
+
+    @Override
+    public boolean canSpawn(World world, int x, int z) {
+        return true;
+    }
+
+    @Override
+    public Location getFixedSpawnLocation(World world, Random random) {
+        final Location spawnLocation = new Location(world, 0.0D, 64.0D, 0.0D);
+        final Location blockLocation = spawnLocation.clone().subtract(0D, 1D, 0D);
+        blockLocation.getBlock().setType(Material.BEDROCK);
+        return spawnLocation;
+    }
+
+}
